requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
psutil==6.0.0
PyJWT==2.9.0
pytest==8.3.3
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
tqdm==4.66.5
twilio==9.3.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.13.1
